# Make the http requestListener the "fastify object"