--- a/package.json
+++ b/package.json
@@ -89,6 +89,7 @@
     "snazzy": "^7.0.0",
     "split2": "^2.2.0",
     "standard": "^10.0.3",
+    "supertest": "^3.0.0",
     "tap": "^11.0.0",
     "then-sleep": "^1.0.1",
     "typescript": "^2.6.2",