@mcollina 2018-01-23
I am -1 because it compromises the bootstrap lifecycle.
I think we should be exporting a function at all, if this is needed, we can
return it with a callback/promise after ready() has been called.

(on a plane with bad connectivity)

Il giorno mar 23 gen 2018 alle ore 18:30 Tommaso Allevi <
notifications@github.com> ha scritto:

> *@allevo* approved this pull request.
>
> LGTM
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/fastify/fastify/pull/705#pullrequestreview-90912737>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AADL4_0F-Aw1aJ-S4oWpbCYCC7zjnt_Oks5tNhdKgaJpZM4Rp9Jo>
> .
>