@nwoltman 2018-01-23
I'm 👍 with making `require('fastify')()` return a plain object instead of a function.