@delvedor 2018-01-23
Closed by #708.